refactor(desktop): move existing desktop components to /desktop directory
- Moved current home page to /desktop (app/desktop/page)
- Moved all associated components to /desktop/components
- Organized existing desktop-related code into a dedicated directory
