+++ app/desktop/components/DockContainer.tsx
