+++ app/desktop/components/DockIcon.tsx
