+++ app/desktop/components/windows/SettingsWindow/PatternSelector.tsx
