+++ app/desktop/components/windows/types.ts
